# api.yaml
openapi: "3.1.0"
info:
  title: "Backbone REST API"
  description: |
    backbone-rest Backoffice REST Service — Java 21 / Spring Boot 4.1.0.
    All endpoints under /api/v1/* require a valid Bearer JWT token unless
    explicitly marked as public (security: []).
  version: "1.0.0"
servers:
  - url: "https://localhost:8084"
    description: "Local development server (TLS 1.3 — see server.port / server.ssl in application.yml)"
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/contacts/:
    post:
      tags: ["contacts"]
      operationId: "createContact"
      summary: "Create a contact"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactCreateRequest"
      responses:
        "201":
          description: "Contact created"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
        "400":
          description: "Invalid contact payload"
        "404":
          description: "Person or application not found"
        "409":
          description: "This person already has a contact with the same value (phone number, email or messaging account) for this application"
        "406":
          description: "Contact limit reached for this person"
  /api/v1/contacts/{contactId}:
    get:
      tags: ["contacts"]
      operationId: "findContactById"
      summary: "Find contact by ID"
      parameters:
        - name: "contactId"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
      responses:
        "200":
          description: "Contact found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
        "404":
          description: "Contact not found"
    put:
      tags: ["contacts"]
      operationId: "updateContact"
      summary: "Update a contact"
      parameters:
        - name: "contactId"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactCreateRequest"
      responses:
        "200":
          description: "Contact updated"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
        "400":
          description: "Invalid payload"
        "404":
          description: "Contact not found"
        "409":
          description: "This person already has a contact with the same value (phone number, email or messaging account) for this application"
    delete:
      tags: ["contacts"]
      operationId: "deleteContact"
      summary: "Delete a contact"
      parameters:
        - name: "contactId"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
      responses:
        "200":
          description: "Contact deleted"
          content:
            application/json:
              schema:
                type: "string"
        "404":
          description: "Contact not found"
  /api/v1/contacts/list/{contactIds}:
    get:
      tags: ["contacts"]
      operationId: "listContactsByIds"
      summary: "List contacts by IDs"
      parameters:
        - name: "contactIds"
          in: "path"
          required: true
          schema:
            type: "array"
            items:
              type: "string"
              format: "uuid"
      responses:
        "200":
          description: "Contact list returned"
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/Contact"
        "404":
          description: "No contacts found"
  /api/v1/contacts/person/{personId}:
    get:
      tags: ["contacts"]
      operationId: "listContactsByPersonId"
      summary: "List contacts by person ID"
      parameters:
        - name: "personId"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
      responses:
        "200":
          description: "Contact list returned"
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/Contact"
        "404":
          description: "Person not found"
...
  /api/v1/managed-clients:
    post:
      tags: [managed-clients]
      summary: Register a new managed client
      description: Returns the clientSecret exactly once in the response body.
      operationId: registerManagedClient
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ManagedClientCreateRequest'
      responses:
        '201':
          description: Client registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagedClientCreateResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagedClientErrorResponse'
        '409':
          description: Duplicate client name for application
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagedClientErrorResponse'
    get:
      tags: [managed-clients]
      summary: List managed clients
      operationId: listManagedClients
      security:
        - bearerAuth: []
      parameters:
        - name: applicationId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: active
          in: query
          required: false
          schema:
            type: boolean
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Client list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ManagedClientTO'
        '204':
          description: No clients found
  /api/v1/managed-clients/token:
    post:
      tags: [managed-clients]
      summary: Issue an M2M access token
      description: Public endpoint. Validates clientId, clientSecret, and requested scopes.
      operationId: issueManagedClientToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ManagedClientTokenRequest'
      responses:
        '200':
          description: Token issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagedClientTokenResponse'
        '400':
          description: Invalid scope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagedClientErrorResponse'
        '401':
          description: Invalid credentials or inactive client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagedClientErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagedClientErrorResponse'
  /api/v1/managed-clients/introspect:
    post:
      tags: [managed-clients]
      summary: Introspect an M2M access token
      description: Public endpoint. Always returns HTTP 200. active=false for invalid, expired, or revoked tokens.
      operationId: introspectManagedClientToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
                  description: The raw M2M access token to introspect
      responses:
        '200':
          description: Introspection result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagedClientTokenIntrospectResponse'
  /api/v1/managed-clients/{clientId}:
    get:
      tags: [managed-clients]
      summary: Get a managed client by ID
      operationId: getManagedClient
      security:
        - bearerAuth: []
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Client found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagedClientTO'
        '404':
          description: Client not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagedClientErrorResponse'
    put:
      tags: [managed-clients]
      summary: Update a managed client
      operationId: updateManagedClient
      security:
        - bearerAuth: []
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ManagedClientUpdateRequest'
      responses:
        '200':
          description: Client updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagedClientTO'
        '404':
          description: Client not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagedClientErrorResponse'
    delete:
      tags: [managed-clients]
      summary: Delete a managed client
      description: Revokes all active tokens before deletion.
      operationId: deleteManagedClient
      security:
        - bearerAuth: []
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Client deleted
        '404':
          description: Client not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagedClientErrorResponse'
  /api/v1/managed-clients/{clientId}/rotate-secret:
    post:
      tags: [managed-clients]
      summary: Rotate the secret for a managed client
      description: Returns the new plaintext clientSecret once. Grace period allows the old secret temporarily.
      operationId: rotateManagedClientSecret
      security:
        - bearerAuth: []
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Secret rotated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagedClientSecretRotateResponse'
        '404':
          description: Client not found or inactive
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagedClientErrorResponse'
  /api/v1/managed-clients/{clientId}/tokens:
    delete:
      tags: [managed-clients]
      summary: Revoke all active tokens for a client
      operationId: revokeAllClientTokens
      security:
        - bearerAuth: []
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: All tokens revoked
        '404':
          description: Client not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagedClientErrorResponse'

  # ── Service Type ─────────────────────────────────────────────────────────────
  /api/v1/service-types:
    get:
      tags: ["service-type"]
      operationId: "listAllServiceTypes"
      summary: "List all service types"
      description: "Returns every registered service type."
      responses:
        "200":
          description: "Service type list returned"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ServiceType"
        "404":
          description: "No service types found"
  /api/v1/service-types/{active}:
    get:
      tags: ["service-type"]
      operationId: "listServiceTypesByStatus"
      summary: "List service types by active status"
      description: "Returns service types filtered by active flag."
      parameters:
        - name: active
          in: path
          required: true
          schema:
            type: boolean
      responses:
        "200":
          description: "Service type list returned"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ServiceType"
        "404":
          description: "No service types found"
  /api/v1/service-types/find/{serviceTypeId}:
    get:
      tags: ["service-type"]
      operationId: "findServiceTypeById"
      summary: "Find service type by ID"
      description: "Returns a single service type by its UUID."
      parameters:
        - name: serviceTypeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: "Service type found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceType"
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "Service type not found"
  /api/v1/service-types/:
    post:
      tags: ["service-type"]
      operationId: "createServiceType"
      summary: "Create a service type"
      description: "Creates a new service type."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceTypeRequest"
      responses:
        "201":
          description: "Service type created successfully"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceType"
        "400":
          description: "Invalid request body"
        "409":
          description: "Service type name already in use"
  /api/v1/service-types/{serviceTypeId}:
    put:
      tags: ["service-type"]
      operationId: "updateServiceType"
      summary: "Update a service type"
      description: "Updates an existing service type by its UUID."
      parameters:
        - name: serviceTypeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceTypeRequest"
      responses:
        "202":
          description: "Service type updated"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceType"
        "400":
          description: "Invalid request"
        "404":
          description: "Service type not found"

  # ── Report ──────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: "http"
      scheme: "bearer"
      bearerFormat: "JWT"
      description: "Standard Bearer JWT token in the Authorization header"
    sessionToken:
      type: "apiKey"
      in: "header"
      name: "session-token"
      description: "Session JWT issued by the backbone-rest session service"
  schemas:
...
    ContactType:
      type: "object"
      properties:
        id:
          type: "string"
          format: "uuid"
        name:
          type: "string"
        description:
          type: ["string", "null"]
        active:
          type: "boolean"
...
    Contact:
      type: "object"
      properties:
        id:
          type: "string"
          format: "uuid"
        content:
          type: "string"
        contactType:
          $ref: "#/components/schemas/ContactType"
        personId:
          type: "string"
          format: "uuid"
        applicationId:
          type: "string"
          format: "uuid"
        active:
          type: "boolean"
    ContactCreateRequest:
      type: "object"
      description: "Wire shape for creating/updating a contact — references the contact type by id (contentTypeId) rather than embedding a full ContactType object."
      properties:
        id:
          type: "string"
          format: "uuid"
        content:
          type: "string"
        contentTypeId:
          type: "string"
          format: "uuid"
        personId:
          type: "string"
          format: "uuid"
        applicationId:
          type: "string"
          format: "uuid"
        active:
          type: "boolean"
...
    ManagedClientCreateRequest:
      type: object
      required: [name, applicationId, scopes]
      properties:
        name:
          type: string
          maxLength: 128
        description:
          type: string
          maxLength: 512
        applicationId:
          type: string
          format: uuid
        scopes:
          type: array
          items:
            type: string
          minItems: 1
        active:
          type: boolean
          default: true
    ManagedClientCreateResponse:
      type: object
      properties:
        clientId:
          type: string
          format: uuid
        clientSecret:
          type: string
          description: Plaintext secret — returned exactly once. Store immediately and securely.
        name:
          type: string
        applicationId:
          type: string
          format: uuid
        scopes:
          type: array
          items:
            type: string
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
    ManagedClientTO:
      type: object
      description: Read-only view of a managed client. Never contains secret material.
      properties:
        clientId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        applicationId:
          type: string
          format: uuid
        scopes:
          type: array
          items:
            type: string
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        lastUpdatedAt:
          type: string
          format: date-time
        secretLastRotatedAt:
          type: string
          format: date-time
    ManagedClientUpdateRequest:
      type: object
      description: All fields optional — only provided fields are updated.
      properties:
        name:
          type: string
          maxLength: 128
        description:
          type: string
          maxLength: 512
        scopes:
          type: array
          items:
            type: string
        active:
          type: boolean
    ManagedClientTokenRequest:
      type: object
      required: [clientId, clientSecret, scopes]
      properties:
        clientId:
          type: string
          format: uuid
        clientSecret:
          type: string
        scopes:
          type: array
          items:
            type: string
          minItems: 1
    ManagedClientTokenResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: Signed RS256 M2M JWT
        tokenType:
          type: string
          example: Bearer
        expiresIn:
          type: integer
          description: Token TTL in seconds
        scopes:
          type: array
          items:
            type: string
        issuedAt:
          type: string
          format: date-time
    ManagedClientSecretRotateResponse:
      type: object
      properties:
        clientId:
          type: string
          format: uuid
        clientSecret:
          type: string
          description: New plaintext secret — returned exactly once. Store immediately and securely.
        gracePeriodSeconds:
          type: integer
          description: Seconds the old secret remains valid
        rotatedAt:
          type: string
          format: date-time
    ManagedClientTokenIntrospectResponse:
      type: object
      required: [active]
      properties:
        active:
          type: boolean
        clientId:
          type: string
        clientName:
          type: string
        scopes:
          type: array
          items:
            type: string
        issuer:
          type: string
        exp:
          type: integer
          format: int64
        iat:
          type: integer
          format: int64
        jti:
          type: string
    ManagedClientErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: invalid_client
        errorDescription:
          type: string
        clientId:
          type: string
    ServiceType:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 128
        description:
          type: string
          maxLength: 512
        active:
          type: boolean
    ServiceTypeRequest:
      type: object
      properties:
        serviceType:
          $ref: "#/components/schemas/ServiceType"

    # ── Addresses ───────────────────────────────────────────────────────────
...