# server/routes/service-types.routes.js
'use strict';

const { Router } = require('express');
const { proxyToBackbone } = require('../shared/proxy');

const router = Router();

router.post('/', (req, res) => proxyToBackbone(req, res, '/api/v1/service-types/'));
router.get('/list-all', (req, res) => proxyToBackbone(req, res, '/api/v1/service-types/true'));
router.get('/find/:serviceTypeId', (req, res) =>
  proxyToBackbone(req, res, `/api/v1/service-types/find/${req.params['serviceTypeId']}`)
);
router.get('/', (req, res) => proxyToBackbone(req, res, '/api/v1/service-types'));
router.get('/:active', (req, res) =>
  proxyToBackbone(req, res, `/api/v1/service-types/${req.params['active']}`)
);
router.put('/:serviceTypeId', (req, res) =>
  proxyToBackbone(req, res, `/api/v1/service-types/${req.params['serviceTypeId']}`)
);

module.exports = router;